requirements.txt
click==8.1.3
DateTime=5.0
geopy==2.3.0
meteostat==1.6.5
numpy==1.24.1
pandas==1.5.3
python-decouple==3.7
prettytable==3.6.0
requests==2.28.2